requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-filter<24
celery>=5.5
redis>=5.0